requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
anthropic==0.3.11
pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
litellm==1.0.0
-e .
